clean

+++ RemoteSubScritps/kill.sh
@@ -1,16 +1,14 @@
 #!/bin/bash
 
-#input discription
 # $1 ---> port that want to kill (optional)
 
-#variables
 port=$1
 
 DIR=`dirname $0`
 input="$DIR/proc.txt"
 
 declare -i index=0
-#methods
+###############################methods
 initializeArr(){
   index=0
   for var in $line ; do
@@ -18,9 +16,7 @@ initializeArr(){
  		index=$index+1
   done
 }
-
-#main method
-main() {
+readFileAndKill(){
   while IFS= read -r line ; do
     echo $line
     initializeArr
@@ -29,11 +25,12 @@ main() {
     #arr[2] ---> pid of process
     if ! [[ -n $port ]]; then
   	  kill ${arr[2]}
-    elif [[ $port = ${arr[1]} ]]; then
+    elif [[ $port = ${arr[1]} ]] && [[ ${arr[0]} = "server" ]]; then
+      echo ${arr[1]} > $DIR/killed.txt
  	    kill ${arr[2]}
- 	  fi
-  done < $input
+    fi
+  done < $1
 }
+###############################
 
-#
-main
+readFileAndKill $input
