+++ kill.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
 
-#input discription
 # $1 ---> port that want to kill (optional)
 
-#variables
 port=$1
 
 DIR=`dirname $0`
@@ -18,9 +16,7 @@ initializeArr(){
  		index=$index+1
   done
 }
-
-#main method
-main() {
+readFileAndKill(){
   while IFS= read -r line ; do
     echo $line
     initializeArr
@@ -29,10 +25,15 @@ main() {
     #arr[2] ---> pid of process
     if ! [[ -n $port ]]; then
   	  kill ${arr[2]}
-    elif [[ $port = ${arr[1]} ]]; then
+    elif [[ $port = ${arr[1]} ]] && [[ ${arr[0]} = "server" ]]; then
+      echo ${arr[1]} > $DIR/killed.txt
  	    kill ${arr[2]}
- 	  fi
-  done < $input
+    fi
+  done < $1
+}
+#main method
+main(){
+  readFileAndKill $input
 }
 
 #
